move `GeneralPath` to `_types` to modify `Filename/PathOrStr`
